+++ controllers/finish_type.controller.js
@@ -12,9 +12,9 @@ exports.getAll = async (req, res) => {
 };
 
 exports.create = async (req, res) => {
-  const { name } = req.body;
+  const { name, SurfaceId } = req.body;
   try {
-    await FinishType.create({ name, image: req.file.filename  });
+    await FinishType.create({ name, image: req.file.filename,SurfaceId });
     return res
       .status(200)
       .json({ success: true, message: "Project Type created successfully" });
@@ -23,10 +23,28 @@ exports.create = async (req, res) => {
   }
 };
 
+exports.getSpecificFinish = async (req, res) => {
+  const { id } = req.body;
+  try {
+    if (!id) {
+      throw "Surface Id is not sent."
+    }
+    const result = await FinishType.findAll({ where: { SurfaceId: id } });
+    return res
+      .status(200)
+      .json({ success: true, message: result });
+  }
+  catch (err) {
+    return res.status(500).json({ success: false, errors: err });
+  }
+}
 exports.update = async (req, res) => {
   const updateFinishType = req.body;
   const { id } = req.params;
-  if(req.file) {
+  if (!id) {
+    throw "Id is missing or incorrect format";
+  }
+  if (req.file) {
     updateFinishType['image'] = req.file.filename;
     const { image } = await FinishType.find({ where: { id: req.params.id }, raw: true });
     fs.unlinkSync(`${keys.storage}/${image}`);
@@ -49,6 +67,9 @@ exports.delete = async (req, res) => {
   const { image } = await FinishType.find({ where: { id: id }, raw: true });
   fs.unlinkSync(`${keys.storage}/${image}`);
   try {
+    if (!id) {
+      throw "Id is missing or incorrect format";
+    }
     await FinishType.destroy({
       where: {
         id
