+++ controllers/pallet.controller.js
@@ -1,4 +1,18 @@
 const { Pallet, Shades } = require("../models");
+
+
+exports.computeDistance = (curr, original) => {
+    curr = curr.map(Number);
+    let finalDistance = { distance: 9999, id: 0 };
+    for (let i of original) {
+        let distance = Math.sqrt(Math.pow((i['r'] - curr[0]), 2) +
+            Math.pow((i['g'] - curr[1]), 2) +
+            Math.pow((i['b'] - curr[2]), 2));
+        finalDistance = { ...finalDistance['distance'] < distance ? finalDistance : { distance, id: i['id'] } };
+    }
+    return finalDistance['id'];
+};
+
 exports.getAll = async (req, res) => {
     try {
         const result = await Pallet.findAll({
@@ -18,18 +32,6 @@ exports.getAll = async (req, res) => {
                 {
                     model: Shades,
                     as: 'color_4'
-                },
-                {
-                    model: Shades,
-                    as: 'color_5'
-                },
-                {
-                    model: Shades,
-                    as: 'color_6'
-                },
-                {
-                    model: Shades,
-                    as: 'color_7'
                 }
             ]
         });
@@ -40,8 +42,7 @@ exports.getAll = async (req, res) => {
 };
 
 exports.create = async (req, res) => {
-    /*
-        {
+    /*{
             "color1Id":"1",
             "color2Id":"2",
             "color3Id":"3",
@@ -49,10 +50,20 @@ exports.create = async (req, res) => {
             "name":"newPalleta",
             "likes":10,
             "pallete_by":"Amjad"
-        }
-    */
-
+    }*/
     try {
+        const { rgb } = req.body;
+        if (rgb) {
+            const rgbValues = await Shades.findAll({ attributes: ['r', 'g', 'b', 'id'], raw: true });
+            let shadeIds = [];
+            rgb.forEach((item, index) => {
+                shadeIds[index] = exports.computeDistance(item, rgbValues);
+            })
+            shadeIds.forEach((item, index) => {
+                req.body[`color${index + 1}Id`] = item;
+            });
+            console.log(req.body)
+        }
         await Pallet.create(req.body);
         return res
             .status(200)
@@ -74,7 +85,9 @@ exports.update = async (req, res) => {
     }*/
     const updateFinishType = req.body;
     const { id } = req.params;
-
+    if (!id) {
+        throw "Id is missing or incorrect format";
+    }
     try {
         await Pallet.update(updateFinishType, { where: { id } });
         return res
@@ -87,6 +100,9 @@ exports.update = async (req, res) => {
 
 exports.delete = async (req, res) => {
     const { id } = req.params;
+    if (!id) {
+        throw "Id is missing or incorrect format";
+    }
     try {
         await Pallet.destroy({
             where: {
